Reintroduce the ExecuteCases and TimeClassExecution classes, to ease code review of this branch.
